requirements.txt
tqdm==4.38.0
torch==1.2.0
nltk==3.4.5
matplotlib==3.1.1
warmup_scheduler==0.1
numpy==1.17.3
torchvision==0.4.0a0+6b959ee
apex==0.1
scikit_image==0.16.2
TorchSnooper==0.7.1
opencv_python==4.1.0.25
docopt==0.6.2
pandas==0.25.3
Pillow==7.1.2
ranger==0.10
skimage==0.0
tensorboardX==2.0
tensorflow==2.2.0
vocab==0.0.4
